backend/requirements.txt
Django>=4.2.16,<4.3
djangorestframework>=3.15.2,<3.16
# psycopg2-binary==2.9.1
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.27.2,<0.28
django-cors-headers>=4.4.0,<4.5.0
# django-debug-toolbar==4.4.6
python-dotenv==1.0.1
djangorestframework-simplejwt>=5.3.1,<5.4.0
# PyJWT==2.9.0
pillow==10.4.0
channels>=4.1.0,<4.2
daphne>=4.1.2,<4.2
channels-redis>=4.2.0,<4.3
asgiref>=3.8.1,<3.9
async-timeout>=4.0.3,<4.1
# selenium>=4.25.0,<4.26
requests==2.32.3
pyotp==2.9.0
qrcode==8.0
googletrans==4.0.0rc1
# twisted[http2,tls]
# celery[redis]
